Refactor: 인터셉터 등록 순서 변경
